introduce utils module with valid_ip_prefix function
valid_ip_prefix checks if a given input is a valid IP address prefix.
IP address prefix contains an IP with a prefix length set after '/'
and in our use case we **always** need the prefix length.
Unfortunately, ipaddress module can't help us here as it doesn't
require the prefix length to be pass and when it is no passed it
is assumed to be 32, see below

In [6]: i = ipaddress.IPv4Address('10.10.10.10')

In [7]: i.max_prefixlen
Out[7]: 32

In [8]: ipaddress.IPv4Network('10.10.10.10')
Out[8]: IPv4Network('10.10.10.10/32')

In [9]: