use the 'to' option in ip tool to list addresses matching the prefix
We used to run 'ip addr show dev lo' to get a list of IPs assigned to
lookback interface in order to check if a IP prefix is assigned.

We introduce the 'to' option of ip tool to only return the IP that matches
ip prefix associated with the service check. See below:

% ip address show dev lo to 10.52.12.3/32
1: lo: <LOOPBACK,UP,LOWER_UP> mtu 65536 qdisc noqueue state UNKNOWN group default
    inet 10.52.12.3/32 brd 10.52.12.3 scope host lo
       valid_lft forever preferred_lft forever
% ip address show dev lo to 10.52.11.3/32
%