Remove unnecessary catch of Empty exception
We accidentally retrieve an item from the queue in a blocking mode as
we pass 1 as an argument to get() method:-). We also catch Empty
exception which is never raised, as get() method will wait until an
item is available. We pass the correct argument(block) and remove
the try catch block.

This change is the result of code review from @carlesnagy