Fix a regression introduced by 37aa574b1159
In 37aa574b1159551e we simplified the way we write IP prefixes in the
bird configuration by using a lambda to generate the list and join
them with comma and new-line character. But, the last element never
got the newline character and as a result a subsequent parsing of the
bird file skipped that IP prefix.

We fix this by adding a leading newline character on the next write.