Change the way threads pass operations
Threads were passing a tuple to the queue for informing main program
to either add or remove IP prefixes from the bird configuration.
In the tuple we stored the IP prefix, action(add or remove) and name
of the service being checked. That was a bit cumbersome approach.

This commit introduces 3 classes which handle the addition and removal
of IP prefixes. Below is the hierarchy of the classes:
BaseOperation
├──AddOperation
├──DeleteOperation

This change is the result of code review from @carlesnagy