Fix a regression introduced by 168f9aab
168f9aab5df1 changed the way we send data from threads to main program.
We used to send a tuple but now we send an object which carries all the
information we need. In 168f9aab5df1 we forgot to send an object when
healthchecking was disabled, we were sending a tuple to the queue
but the consumer of the queue was waiting an objection, which resulted
to an AttributeError exception and termination of the daemon.