Change the way we handle IP prefixes without check
Previously, we considered the situation where there are IP prefixes in
Bird configuration for which we aren't configured to run check against
them as a fatal state. Thus, we exited main program and aborted
configuration sanitization( -c CLI).

This commit changes this behavior by removing IP prefixes from Bird
configuration which are not configured for health checking.

Configuration sanitization was changed to only emit a message
rather aborting.