Add support for JSON logging to a remote point
This commit extends logging functionality by building JSON blobs and
push them over http to a remote point. It is not enabled by default.

Implementation details:
We redesigned the logger class to push over HTTP log messages to a
remote point. A user of LoggerExt class can call add_central_logging
on the instantiated object to add HTTP capabilities.

The structure of the JSON blob is not configurable (yet).

WARNING: The http POST requests are done in blocking mode which means
that possible long delays will make the health checks to be delayed.
`http_server_timeout` option on [daemon] accepts floating point numbers
as values which are passed to underlying request module as a single
timeout which will be applied to both the connect and the read timeouts.

@charlesnagy reviewed the changes in the LoggerExt class.