Wait a bit before starting again the daemon.
This is a change for the script which we use to launch the daemon
in the development system.

After the addition of JSON logging the termination of the daemon takes
~200ms due to the fact that we don't asynchronously push data over the
network.

In production systems we don't see this behavior as the SysVinit script
waits for the termination of the daemon before it exits thus a restart
never fails.