Store anycast-prefixes.conf under /var/lib/
According to Filesystem Hierarchy Standard version 3.0 /var/lib
directory can be used to hold data that are modified during the run of
the program. Thus, we store anycast-prefixes.conf under
/var/lib/anycast-healthchecker directory.

Having said, we don't change our default setting in order to avoid
breakages in current installations.

Note for packagers: This directory must be exists prior the run
of the daemon and a link must be created in /etc/bird.d directory
to point to the new location of the file.