Implement a more atomic way to update bird conf
We used to write directly to the bird configuration which opens the
door for leaving it in an inconsistent if we fail during the write.

We now write first to a temporary file and then perform a rename, which
is an atomic operation. Because we use os.rename we create the temporary
file in the directory where bird configuration is stored.

We also replace a map/lambda combo with a list comprehension.