Update bird config when IP isn't assigned to lo
In case an IP for a service check is not assigned to loopback interface,
we need to remove it from the bird configuration in order to make sure
it is not advertised by any rooting protocols in some corner cases.

The following sequence of events caused an advertisement to be sent
even when the service was down:

1) Service is UP and its associated IP is assigned to lo interface and
present in bird configuration

2) Server reboots

3) anycast-healthchecker runs 1st check and finds the IP for a service
not assigned to loopback interface. It emits a warning and set the
established state to DOWN

4) IPs are assigned to loopback interface

5) Direct protocol from BIRD adds routes for these IPs in RIB

6) BIRD advertises routes for theses IPs as they are present in
the bird configuration

7) anycast-healthchecker runs 2nd check and finds the IP for the service
assigned to the interface but the application is not available yet.

Application wasn't available because at the startup IP wasn't present
and failed to bind to the socket.

Anycast-healthchekcer sets state to going down 1, on then next check
it sets established state to DOWN. Because previous established state
was DOWN anycast-healthchecker doesn't ask for removal of the IP from
the bird configuration.

With the current design we ask for a removal/addition of IPs from/to
bird configuration only when the established state changes. When there
isn't any change (UP-->check1..N-->UP or DOWN-->check1..N-->DOWN) we
don't take any actions.

In a follow-up commit we will change this to always take an action.
We are also going to introduce a check against the information in the
RIB in order to have a more robust setup.