Work around intermittent ValueErrors
We have been getting intermittent ValueErrors, see here
https://gist.github.com/unixsurfer/67db620d87f667423f6f6e3a04e0bff5
when we check if IP is assigned to loopback interface by running
  /sbin/ip address show dev lo to 10.52.12.2/32

It has happened ~5 times and this code is executed from multiple
threads and every ~10secs on several(~40) production servers for
more than 18months.  It could be a bug in Python or system returns
corrupted data.

As a consequence of the raised exception thread dies and the service
isn't monitored anymore!. In order to avoid this we catch the exception.

While checking if IP is assigned, we get an error unrelated to the IP
which prevents us from knowing if it's assigned, we simply don't know.
Therefore, we return True rather False.

A retry logic inside _ip_assigned could be a more proper solution.