Catch the case where check_cmd can't be killed
When a *check_cmd* doesn't return after *check_timeout* seconds we try
to kill it. But, when the *check_cmd* runs under another user (root)
then we don't have sufficient privileges to terminate it and as
consequence the thread dies with:

Exception in thread servicefoo:
Traceback (most recent call last):
  File "../lib/python3.4/site-packages/anycast_healthchecker/servicecheck.py", line 60, in _run_check
    proc.wait(self.config['check_timeout'])
  File "../lib/python3.4/subprocess.py", line 1564, in wait
    raise TimeoutExpired(self.args, timeout)
subprocess.TimeoutExpired: Command '['sudo', '/foocmd.sh']' timed out after 6 seconds

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "../lib/python3.4/threading.py", line 911, in _bootstrap_inner
    self.run()
  File "../lib/python3.4/site-packages/anycast_healthchecker/servicecheck.py", line 222, in run
    elif self._run_check():
  File "../lib/python3.4/site-packages/anycast_healthchecker/servicecheck.py", line 64, in _run_check
    proc.kill()
  File "../lib/python3.4/subprocess.py", line 1698, in kill
    self.send_signal(signal.SIGKILL)
  File "../lib/python3.4/subprocess.py", line 1688, in send_signal
    os.kill(self.pid, sig)
PermissionError: [Errno 1] Operation not permitted

We now prevent the death of thread by catching PermissionError
exception and emit a warning message.

If we run as root user then we have sufficient privileges to terminate
the *check_cmd*. But, it is highly recommended to users to
run anycast-healthchecker as normal user because running under root
user it opens the door for bigger problems.

The disadvantage of this approach is that we may leave orphan processes
(the invoked check_cmd) on the system. This can be mitigated by
using a *check_cmd* which run the health check with a time limit.

For instance, timeout tool can be used:

    [servicefoo]
    ...
    check_cmd = timeout 5s /servicefoocheck.sh
    ...

If the monitored service is a web or DNS application then curl and dig
tools provide the necessary option to set a maximum time for the operation
to finish. That timeout should be less than the timeout set with
*check_timeout*.