Simplify sanity check on configuration
We used to exit main program upon the 1st failure of sanity check and
this commit changes this behavior. We now raise a ValueError exception
from each function which performs the sanitization. The exception is
caught by the main program and aborts with exit code 1 and some useful
message.

We also fixed some typos in docstrings/comments.

There isn't any functional changes which effect the service checks.