Add option to disable the IP assignment check
An IP prefix associated with a service is usually assigned to an interface.
Thus, we were always checking if that IP was assigned to the interface
and set the state of the service to down if the IP wasn't.

But, there are some use-cases where the IP doesn't need to be assigned
to any interface as the applications which receives the traffic works
in transparent mode and is able to bind to a foreign IP address.

This commit introduces ip_check_disabled option which controls this
behavior. Default value is false, so we don't disable it silently to
current deployments.