Move functions to utils module
The functions for writing the Bird configuration in a temporary file
and to reconfigure Bird are moved to utils module.

These 2 functions will be used in the sanity check during start-up.
That sanity check will remove IP-Prefixes from Bird configuration for
which we don't have a service check associated with them.

This is a pre-work for #3.