Add support for removing IP-Prefixes without a check
If purge_ip_prefixes is set to true, then on start-up we remove
IP-Prefixes from Bird configuration for which we don't have a service
check associated with them. Default value is false in order to be
backward compatible with previous versions, otherwise we may cause
problems on environments where don't expect from us to remove those
IP-prefixes.

NOTE: Those IP-Prefixes are always removed from Bird configuration when
it is updated during the life time of the daemon.

Closes #3