Don't exit when there aren't any service checks
This commit changes the behavior of the daemon when there aren't any
service checks configured. We used to exit the daemon in this case,
but not anymore. The daemon will continue to run and emit a log saying
that there aren't any service checks. The daemon will stay up doing
nothing.

Closes #3