Remove die parameter from get_ip_prefixes_from_bird
It is considered bad design to have a function parameter which changes
the behavior of the function.

get_ip_prefixes_from_bird suffers from that as we use `die` to handle
OSError exception in 2 different ways.

We remove `die` parameter and we make sure the callers of that function
catch OSError exception.