Add support for specifying a single service check
The CLI interface was only allowing users to set a directory with
configuration files for service checks. If a user has only one service
to monitor then he was forced to make a directory and save a file under
that directory.

This commit changes CLI interface by introducing -F, --service-file
option which allows users to set a configuration file for a single
service check::

    anycast-healthchecker --file=anycast-healthchecker.conf -F foo.conf

--dir and --service-file options are mutually exclusive, so they can't
be used at the same time. Furthermore, --service-file option doesn't
have a default value.

This new option is very useful in cases where someone wants to perform
a sanity check against a configuration file for one service check::

    anycast-healthchecker -c -f healthchecker.conf -F dir.d/f.conf

Closes: #7