Keep a history of changes for bird configuration
This commit adds support for keeping a history of changes for Bird
configuration file by copying the file to a directory before we truncate
it with new content. During the startup of the daemon a directory with
the name **history** is created under the directory where Bird
configuration file resides. The daemon has to have sufficient privileges
to create that directory.

This functionality is enabled when **keep_bird_changes** is set to
**true** and **changes_counter** sets number of files to keep.

This functionality is not enabled by default.

Closes: #7