Better handling of symbolic links for bird config
If a user sets *bird_config* or/and *bird6_conf* to a symbolic link
then we must ensure that we create the history directory and temporarily
files are under the directory the link points.

We do that by catching this case when we parse the configuration and
store in *config_file* key the canonical path rather the link.

We also print a message to inform the user that we use the canonical
path rather the link.