Create anycast-prefixes.conf configuration
The anycast-prefixes.conf set with ``bird_conf`` and bird6_conf``
settings were required to be present before we could use them.
This hard requirement doesn't exist anymore.

We now create them upon the 1st launch and populate them with the
correct data. We know the variable name (``bird_variable`` and
``bird6_variable``) and the dummy_ip_prefix per IP version, so we have
all the information we need to create a valid configuration.

Documentation and code are changed to reflect that change.