Remove the logic, which creates parent directories
In 6f4d2c61000da532d6f33b7dc65b25b412065964 we introduced the logic of
creating the parent directory of configuration files set in
**bird_conf** and **bird8_conf** settings.

This approach is problematic as with the default configuration it tries
to create *anycast-healthchecker* directory under /var/lib.
anycast-healthchecker runs as normal user, which doesn't have
privileges to create the directory under /var/lib.

This commit removes this logic and we now expect the administrators to
create this directory after the initial installation.