Fail sanity check when config file and dir don't exist
We used to only print a message when either directory with configuration
files for service checks or a file for a single service check were
missing, which caused the daemon to start and do nothing as we didn't
have any service check to perform.

We now fail during the sanity check and abort when we don't have any
service check(s) to perform.