Make sure we return the prefix length
If a user sets ip_prefix setting with an IP address without prefix
length then we remove it from bird configuration. This happens because
the sanity check for IP prefixes without service checks marks those IP
prefixes for removal due to the fact we compare IP prefixes with
IP addresses, which don't have prefix length.

In order to find which IP prefixes in bird configuration don't have a
service check associated with them, we compare two lists of IP prefixes.
One returned by get_ip_prefixes_from_config, which returns the IP
prefixes as they are configured by the user, and one returned by
get_ip_prefixes_from_bird, which returns the IP prefixes found in the
bird configuration. The former may have an element without the prefix
length but the latter we always have elements with prefix lengths.
The fact that one IP address appears in one list with prefix length but
not in the other list, causes the removal of that IP address from bird
configuration.

Since 0.7.0 version we don't force user to set the prefix length in
ip_prefix setting of service checks. We switched to ipaddress module for
checking the validity of the values passed to ip_prefix setting and that
module sets the default prefix length when it receives a network address
as input without a prefix length.

We documented this behavior with de9b763e, but we skipped to update
get_ip_prefixes_from_config().

We fix the issue by making sure that get_ip_prefixes_from_config returns
the prefix length, regardless if it is set by the user.