Catch the case where pid is higher than allowed
During the startup we check if pidfile exists and if it does then we
check if the process with pid mentioned in the pidfile is still alive.
If that pid is higher than allowed then we crash with:
    anycast-healthchecker -f ./var/etc/anycast-healthchecker.conf -d ./var/etc/anycast-healthchecker.d
    acquired a lock by creating an abstract namespace socket
    Traceback (most recent call last):
    File "foopath/.virtualenvs/python3/bin/anycast-healthchecker", line 10, in <module>
    sys.exit(main())
    File "foopath/.virtualenvs/python3/lib/python3.5/site-packages/anycast_healthchecker/main.py", line 81, in main
        update_pidfile(pidfile)
    File "foopath/.virtualenvs/python3/lib/python3.5/site-packages/anycast_healthchecker/utils.py", line 812, in update_pidfile
        if running(pid):
    File "foopath/.virtualenvs/python3/lib/python3.5/site-packages/anycast_healthchecker/utils.py", line 556, in running
        os.kill(processid, 0)
    OverflowError: signed integer is greater than maximum

We now catch this very rare case and exit normally.

Under normal circumstance this will never happen as we write to pidfile
and we also remove it upon shutdown. But, a malicious user could create
the pidfile and put an integer greater than it is allowed.