Dynamically set width in log formatter
We wanted to easier reading the logs and when we have a lot of service
checks and the name of those services have different length.
In order to do that we used to hard code the width of the 'service name'
in logs to 32 chars, so we can vertically align the service names with
the log message, which comes after the 'service name'.

We now dynamically set the width by finding the service name with longest
string and set the width to be that plus one more character.