Exit main program when a thread dies
The main program isn't aware when a thread dies due to an unhandled
exception or thread crash, and without the thread running we don't
perform any service checks. Since the main program isn't aware of that
situation, it doesn't try to recover the crashed thread.

In order to mitigate this, we have a nagios check under contrib/nagios
directory, which checks if any of the thread has crashed. It also checks
if the main program has crashed.

The above mitigation is an suboptimal solution as it exposes the
internals of anycast-healthchecker to an external entity and if we
want to make a change then we have to update that external entity.

But, there is simple way for main program to recover from this
situation. We can simply exit main program and let systemd restarts us.
If we exit with non-zero code and systemd service option 'Restart' is
set to 'on-failure' for our systemd service, our default systemd
service configuration under contrib/systemd is configured like this,
then systemd will notice that we have exited with non-zero code and
it will trigger a restart.

How do we know if a thread has died? This is a bit tricky to get it
right. But, for now we catch all possible exception that may be raised
during the life time of a thread and inform main program about it.

We should also implement a healthiness check from the main program for
each thread by having hello messages sent by threads and let main program
exit if we miss X number of consecutive message from a thread. That will
be a more robust a complete solution, but it is tricky to get it right.