Ensure correct variable name in Bird configuration
Upon start up we don't check if the variable name set in Bird
configuration is different with one set in bird_variable option. This
can lead to issues with configuration management tools(Puppet and etc).
We will eventually fix this the first time we modify IP prefixes
as the result of a service check being failed for an IP prefix which is
present in the Bird configuration. But, this is a suboptimal solution
because after the startup we may not modify the file at all, due to the
fact none of the service checks failed or succeed.

Thus, we now check during the startup if variable name in Bird
configuration is the same with one set in bird_variable option. If
it doesn't match then we add the correct one. We do the same if we don't
find the variable in the Bird configuration at all.