Add support for custom Bird reconfigure per service
We can now run a custom command set per service to trigger a
reconfiguration on Bird upon a change in the status of the service.
Users can set the service option **custom_bird_reconfigure_cmd** to a
script or command and we run that instead the one set with
**bird_reconfigure_cmd** in Daemon section. When we run the custom
command we pass a string as an argument, which is either "up" or "down".

Close #18