Exit if configuration file passed with '-f' doesn't exist
We pass to read() method of ConfigParser object a list of configuration
file and if a file doesn't exist the aforementioned method will simply
ignore it and proceed.

This may confuse people when they pass a configuration file which
doesn't exist as we, in that case, use the defaults for all settings in
the daemon section. In order to avoid this situation we exit main
program when we can't read configuration file set with '-f' argument.