Pass the exit code of check_cmd to service_state metric
From an observability point of view, it is more useful if we pass the
exit code of the service check_cmd to the `service_state` Prometheus
metric instead setting the `exitcode` metric. The former metric conveys
the established state and not the intermediate state.

If the `chec_cmd` times out, we set `127` as the value of the `service_state`
metric.

This commit does not change the logic to determine the state of the service
in any way.